mount /var/lib/docker so docker magically works
this is a hack; we should ultimately mount it as "/scratch" or something
generic.

the point of this is to have a non-overlay mount point so that docker
can make its own aufs or overlay mounts.

the volume will be GCed when the container goes away. 'life' makes this
kind of thing super easy!

concourse/concourse#1045