make all resources but 'docker-image' unprivileged
this makes it so that the vast majority of containers managed by
concourse are unprivileged, so they can all benefit from their images
and input artifacts/volumes being unprivileged.

ultimately this should be done by the base resource type being listed as
privileged, so that we're not special-casing by name.

we'll probably also need 'privileged: true' in the 'resource_types'
section of the pipeline.

concourse/concourse#1045