move evaluating of secrets to build step
The pulling from vault originally happened right
before the build plan was created but now we moved
it down to when the build step runs in exec. We
also added a new source object in the creds package
which will set up the pulling from vault and
templating out the variables.

concourse/concourse#291

Signed-off-by: Topher Bullock <cbullock@pivotal.io>